# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PalPanel API
  version: 1.0.0
  description: Stable HTTP contract for the PalPanel 1.x management API.
servers:
  - url: /api
security:
  - sessionCookie: []
  - developmentKey: []
...
paths:
...
  /mods/import/inspect:
    post:
      operationId: inspectModImport
      tags: [content]
      x-palpanel-permission: 'mods:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModImportInspectRequest'}
          multipart/form-data:
            schema: {$ref: '#/components/schemas/ModImportUploadRequest'}
      responses: {'200': *importInspection, '400': *error, '401': *error, '403': *error, '413': *error, '502': *error, '500': *error}
  /mods/import/inspect/{id}/select:
    parameters: [*idParameter]
    post:
      operationId: selectModImportCandidate
      tags: [content]
      x-palpanel-permission: 'mods:write'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModImportSelectRequest'}
      responses: {'200': *importInspection, '400': *error, '401': *error, '403': *error, '404': *error, '409': *error, '410': *error, '413': *error, '502': *error, '500': *error}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: palpanel_session
      description: HttpOnly SameSite=Lax session cookie created by registration or login.
    developmentKey:
      type: http
      scheme: bearer
      bearerFormat: ppk_...
      description: 'Revocable development key supplied as Authorization: Bearer ppk_...'
  schemas:
...
    ModImportInspectRequest:
      type: object
      required: [source]
      properties:
        source: {type: string}
    ModImportUploadRequest:
      type: object
      required: [file]
      properties:
        file: {type: string, format: binary}
    ModImportSelectRequest:
      type: object
      required: [candidate_id]
      properties:
        candidate_id: {type: string}
...